consts fixes. directory structure refactored. eventMixin.trigger() made private
