+++ src/models/profile/Profile.js
@@ -1,4 +1,4 @@
-import {Model} from './Model.js';
+import {Model} from '../Model.js';
 
 /**
  * Profile model
