+++ src/models/tag/Tag.js
@@ -1,4 +1,4 @@
-import {Model} from './Model.js';
+import {Model} from '../Model.js';
 
 /**
  * Tag model
